requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.3.0
pydantic==2.7.1
